+++ v2/file/file.go
@@ -2,9 +2,9 @@
 package file
 
 import (
+	"errors"
+	"io/fs"
 	"os"
-	"path/filepath"
-	"strings"
 )
 
 // Put 将数据存入文件.
@@ -15,23 +15,14 @@ func Put(data []byte, to string) error {
 // Exists 判断文件是否存在.
 func Exists(fileToCheck string) bool {
 	_, err := os.Stat(fileToCheck)
-	return !os.IsNotExist(err)
-}
-
-// FileNameWithoutExtension 去除文件扩展名.
-func FileNameWithoutExtension(fileName string) string {
-	return strings.TrimSuffix(fileName, filepath.Ext(fileName))
+	return !errors.Is(err, fs.ErrNotExist)
 }
 
 // CreateDirIfNotExists 创建目录（若不存在）.
-func CreateDirIfNotExists(dirname string, perm ...os.FileMode) error {
+func CreateDirIfNotExists(dirname string) error {
 	if _, err := os.Stat(dirname); err != nil {
-		if os.IsNotExist(err) {
-			p := os.ModePerm
-			if len(perm) > 0 {
-				p = perm[0]
-			}
-			return os.MkdirAll(dirname, p)
+		if errors.Is(err, fs.ErrNotExist) {
+			return os.MkdirAll(dirname, os.ModePerm)
 		}
 		return err
 	}